requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-dotenv==1.0.1

langchain==0.3.23
langchain-community==0.3.21
langchain-google-genai==2.1.2
langchain-chroma==0.2.2
chromadb==0.6.3
